Refactor server components out of index
